Overhauled OwnerAuthorizationHandler to fix a potential security flaw in the previous design.
With the flexibility of the list of owner parameter names to search for, an attacker could pass their valid user id as a parameter accepted by the OwnerAuthorizationHandler and pass a different user id for the actual controller, defeating the authorization check.
Additionally, the two sources of parameters (route and headers) could cause a similar interference.
While this depends on the ordering of name candidates, and could be fixed by configuration, the issue can be much more reliably prevented, by not providing such a candidate list in the first place.
Instead, this switches from providing one "Owner" policy to providing both a "RouteOwnerUserId" and a "HeaderOwnerUserId" policy, both of which only allow a single kind of parameter (either route OR headers) and only work with a parameter named "UserId".
However, note that both underlying parameter management mechanisms use maps that ignore the case.
Therefore, e.g. "userId", "userid", and "USERID" should also work.