Flush log writer after each message to ensure that all messages are actually recorded to the log files.
While this incurrs some performance costs, it is unavoidable for correctness because without it log entries on rarely used writers might never be written to the file and keep stuck in the buffers at least until shutdown.
This is especially problematic if there are sinks for rare but important events, which are then still in-memory only.