Removed DbContextExtensions as the type parameter inference doesn't work correctly for the instance overload of LoadKeyValuePropertiesAsync.
Deriving TDefinition through the constraint of TPropInstance seems to be a layer too much for deduction.
Explicitly specifying type arguments fixes the problem, but makes the call quite verbose.
As the call was already not that much shorter than calling Entry, Collection, and LoadAsync, the extension methods make not much sense anymore.