Implemented workaround for a problem with the key-value properties in EF Core change tracking.
Background:
When calling SaveChanges after adding a property to a property-owning entity,
the change tracker kicks in and determines the changes made to the tracked entities.
With the previous version of the key-value properties, the property objects (both definitions and instances)
generated a GUID primary key when the objects were created.
However, as the property entities are non-owned (as another workaround because otherwise instances couldn't reference instances)
and already have a primary key, the changes tracker assumes them to be existing entities,
that it didn't know about and just detects the supposedly changed foreign key for Owner.
As a consequence, it generates an UPDATE command instead of an INSERT command.
Because this update references a non-existing primary key, it then affects 0 rows and EF Core misinterprets this as a DbUpdateConcurrencyException.

To work around this issue, the property objects now get Guid.Empty as their initial Id.
The change tracker detects this, tracks them as added and assigns them Guid Ids.
In contrast to sequence ids this is even done before going to the database, immediately when the change tracker is triggered and adds these objects.

As a consequence of Ids potentially being Guid.Empty, property validataion can no longer rely on them.
Instead, the searched objects are now looked up using object identity with a fallback on name comparison.
The latter is done to still work in cases where strict identity mapping is not provided.