Breaking change for SGL.Utilities.Crypto: Use terminology of message key pair / message public key instead of sender key pair / sender public key to avoid confusion between the message-specific, ephemeral key for ECDH and the permanent public key of the sender.
For the use-case in SGL.Analytics this was not a problem as the senders don't have associated key pairs there, and thus, only the ephemeral keys are used.
For use with SGL.Community, this changes, as the encrypted messages are exchanged between multiple users which each have their own permanent public key.
Thus this breaking change is the lesser evil, compared to confusing code and possibly resulting bugs.