For pre-release versions, increment the last digit of the number.
This corrects for a difference between git describe output and SemVer:
Suffixed versions in git describe are *after* the tag with the version number,
suffixed versions in SemVer are candidates for the version and thus before the release version.
This is imperfect, as the automation can't know which digit the next version will increment.
So, pre-releases for a breaking change may be misattributed as pre-releases for a revision update.
However, this is still the lesser evil compared to backward pre-releases.