Fixed exception handling of MapBufferedAsync.
Ensure that when an exception is thrown by one of the mapping tasks,
the yield buffer ist correctly drained of already completed results,
and the pending tasks in the map buffer are cleanly terminated by awaiting them,
instead of just dropping the task objects.
To preserve ordering semantics, the results of the tasks after the one that threw the exception
are dropped, but if more tasks throw exceptions, all caught exceptions are bundled into an AggregateException,
unless all exceptions are task cancellations, then we forward our own OperationCanceledException.